+++ asreview/webapp/src/PreReviewComponents/index.js
@@ -3,7 +3,6 @@ export { default as Help, useHelp } from "../PreReviewComponents/Help";
 export { default as LabeledItems } from "../PreReviewComponents/LabeledItems";
 export { default as ListItemPaper } from "../PreReviewComponents/ListItemPaper";
 export { default as PaperCard } from "../PreReviewComponents/PaperCard";
-export { default as PluginDataset } from "../PreReviewComponents/PluginDataset";
 export { default as PreReviewZone } from "../PreReviewComponents/PreReviewZone";
 export { default as PriorKnowledge } from "../PreReviewComponents/PriorKnowledge";
 export { default as PriorKnowledgeRandom } from "../PreReviewComponents/PriorKnowledgeRandom";
@@ -13,9 +12,8 @@ export { default as ProjectInfo } from "../PreReviewComponents/ProjectInfo";
 export { default as ProjectModeSelect } from "../PreReviewComponents/ProjectModeSelect";
 export { default as ProjectPageOLD } from "../PreReviewComponents/ProjectPageOLD";
 export { default as ProjectUpload } from "../PreReviewComponents/ProjectUpload";
-export { default as ProjectUploadBenchmarkDatasets } from "../PreReviewComponents/ProjectUploadBenchmarkDatasets";
+export { default as ProjectUploadDatasets } from "../PreReviewComponents/ProjectUploadDatasets";
 export { default as ProjectUploadFile } from "../PreReviewComponents/ProjectUploadFile";
-export { default as ProjectUploadPluginDatasets } from "../PreReviewComponents/ProjectUploadPluginDatasets";
 export { default as ProjectUploadURL } from "../PreReviewComponents/ProjectUploadURL";
 export { default as SearchResult } from "../PreReviewComponents/SearchResult";
 export { default as StartReview } from "../PreReviewComponents/StartReview";
